rest tutorial with jwt-extended

+++ resources/item.py
@@ -1,5 +1,8 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required, current_identity
+from flask_jwt_extended import (
+    jwt_required, get_jwt_claims,
+    jwt_optional, get_jwt_identity, fresh_jwt_required
+)
 from models.item import ItemModel
 
 
@@ -16,7 +19,7 @@ class Item(Resource):
                         help="Every item needs a store id."
                         )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
 
         item = ItemModel.find_by_name(name)
@@ -24,6 +27,7 @@ class Item(Resource):
             return item.json()
         return {'message': "Item not found."}, 404
 
+    @fresh_jwt_required
     def post(self, name,):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -39,7 +43,11 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': "Admin privilege required."}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -63,5 +71,15 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        user_info = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_info:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': "More data available if you log in."
+        }, 200
+
+
